telegram_bot/requirements.txt
aiogram
asyncpg
python-dotenv
yookassa

